requirements.txt
scrapy>=2.11.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
